LayoutTests/platform/glib/media/media-source/media-source-changetype-second-init-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
SourceBuffer.changeType is undefined
END OF TEST
